refactor query editor to pass sections
