+++ app/client/src/components/editorComponents/ActionRightPane/index.tsx
@@ -1,10 +1,8 @@
-import React, { useContext, useEffect, useMemo, useRef, useState } from "react";
+import React, { useContext, useEffect, useMemo, useRef } from "react";
 import styled from "styled-components";
-import { Collapse, Classes as BPClasses } from "@blueprintjs/core";
-import { Classes, getTypographyByKey } from "design-system-old";
-import { Divider, Icon, Text } from "design-system";
+import { getTypographyByKey } from "design-system-old";
+import { Divider } from "design-system";
 import SuggestedWidgets from "./SuggestedWidgets";
-import type { ReactNode, MutableRefObject } from "react";
 import { useParams } from "react-router";
 import { useDispatch, useSelector } from "react-redux";
 import { getWidgets } from "sagas/selectors";
@@ -44,7 +42,6 @@ import {
 } from "utils/storage";
 import { SCHEMA_SECTION_ID } from "entities/Action";
 import { getCurrentUser } from "selectors/usersSelectors";
-import { Tooltip } from "design-system";
 import { ASSETS_CDN_URL } from "constants/ThirdPartyConstants";
 import { FEATURE_WALKTHROUGH_KEYS } from "constants/WalkthroughConstants";
 import { getAssetUrl } from "@appsmith/utils/airgapHelpers";
@@ -52,6 +49,9 @@ import { useFeatureFlag } from "utils/hooks/useFeatureFlag";
 import { FEATURE_FLAG } from "@appsmith/entities/FeatureFlag";
 import { getHasManagePagePermission } from "@appsmith/utils/BusinessFeatures/permissionPageHelpers";
 import { DatasourceStructureContext } from "entities/Datasource";
+import Collapsible, {
+  DisabledCollapsible,
+} from "components/common/Collapsible";
 
 const SCHEMA_GUIDE_GIF = `${ASSETS_CDN_URL}/schema.gif`;
 
@@ -104,43 +104,6 @@ const SideBar = styled.div`
   }
 `;
 
-const Label = styled.span`
-  cursor: pointer;
-`;
-
-const CollapsibleWrapper = styled.div<{
-  isOpen: boolean;
-  isDisabled?: boolean;
-}>`
-  display: flex;
-  flex-direction: column;
-  overflow: hidden;
-  ${(props) => !!props.isDisabled && `opacity: 0.6`};
-
-  &&&&&& .${BPClasses.COLLAPSE} {
-    flex-grow: 1;
-    overflow-y: auto !important;
-  }
-
-  .${BPClasses.COLLAPSE_BODY} {
-    padding-top: ${(props) => props.theme.spaces[3]}px;
-    height: 100%;
-  }
-
-  & > .icon-text:first-child {
-    color: var(--ads-v2-color-fg);
-    ${getTypographyByKey("h4")}
-    cursor: pointer;
-    .${Classes.ICON} {
-      ${(props) => !props.isOpen && `transform: rotate(-90deg);`}
-    }
-
-    .label {
-      padding-left: ${(props) => props.theme.spaces[1] + 1}px;
-    }
-  }
-`;
-
 const Placeholder = styled.div`
   display: flex;
   justify-content: center;
@@ -170,89 +133,14 @@ const CollapsibleSection = styled.div<{ height: string; marginTop?: number }>`
   }
 `;
 
-interface CollapsibleProps {
-  expand?: boolean;
-  children: ReactNode;
-  label: string;
-  CustomLabelComponent?: (props: any) => JSX.Element;
-  isDisabled?: boolean;
-  datasourceId?: string;
-  containerRef?: MutableRefObject<HTMLDivElement | null>;
-}
+const SectionWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+`;
 
-interface DisabledCollapsibleProps {
-  label: string;
-  tooltipLabel?: string;
-}
-
-export function Collapsible({
-  children,
-  containerRef,
-  CustomLabelComponent,
-  datasourceId,
-  expand = true,
-  label,
-}: CollapsibleProps) {
-  const [isOpen, setIsOpen] = useState(!!expand);
-
-  const handleCollapse = (openStatus: boolean) => {
-    if (containerRef?.current) {
-      if (openStatus) {
-        containerRef.current.style.height = "";
-      } else {
-        containerRef.current.style.height = "auto";
-      }
-    }
-    setIsOpen(openStatus);
-  };
-
-  useEffect(() => {
-    handleCollapse(expand);
-  }, [expand]);
-
-  return (
-    <CollapsibleWrapper isOpen={isOpen}>
-      <Label className="icon-text" onClick={() => handleCollapse(!isOpen)}>
-        <Icon
-          className="collapsible-icon"
-          name={isOpen ? "down-arrow" : "arrow-right-s-line"}
-          size="lg"
-        />
-        {!!CustomLabelComponent ? (
-          <CustomLabelComponent
-            datasourceId={datasourceId}
-            onRefreshCallback={() => handleCollapse(true)}
-          />
-        ) : (
-          <Text className="label" kind="heading-xs">
-            {label}
-          </Text>
-        )}
-      </Label>
-      <Collapse isOpen={isOpen} keepChildrenMounted>
-        {children}
-      </Collapse>
-    </CollapsibleWrapper>
-  );
-}
-
-export function DisabledCollapsible({
-  label,
-  tooltipLabel = "",
-}: DisabledCollapsibleProps) {
-  return (
-    <Tooltip content={tooltipLabel}>
-      <CollapsibleWrapper isDisabled isOpen={false}>
-        <Label className="icon-text">
-          <Icon name="arrow-right-s-line" size="lg" />
-          <Text className="label" kind="heading-xs">
-            {label}
-          </Text>
-        </Label>
-      </CollapsibleWrapper>
-    </Tooltip>
-  );
-}
+const CollapsibleGroup = styled.div`
+  flex: 1;
+`;
 
 export function useEntityDependencies(actionName: string) {
   const deps = useSelector((state: AppState) => state.evaluations.dependencies);
@@ -282,6 +170,7 @@ function ActionSidebar({
   hasConnections,
   hasResponse,
   pluginId,
+  sections,
   suggestedWidgets,
 }: {
   actionName: string;
@@ -291,6 +180,7 @@ function ActionSidebar({
   datasourceId: string;
   pluginId: string;
   context: DatasourceStructureContext;
+  sections?: React.ReactNode;
   actionRightPaneBackLink: React.ReactNode;
 }) {
   const dispatch = useDispatch();
@@ -413,53 +303,57 @@ function ActionSidebar({
   return (
     <SideBar>
       {actionRightPaneBackLink}
+      <SectionWrapper>
+        {sections && <CollapsibleGroup>{sections}</CollapsibleGroup>}
+        <CollapsibleGroup>
+          {showSchema && (
+            <CollapsibleSection
+              height={
+                datasourceStructure?.tables?.length && !isLoadingSchema
+                  ? "50%"
+                  : "auto"
+              }
+              id={SCHEMA_SECTION_ID}
+              ref={schemaRef}
+            >
+              <Collapsible
+                CustomLabelComponent={DatasourceStructureHeader}
+                containerRef={schemaRef}
+                datasourceId={datasourceId}
+                expand={!showSuggestedWidgets}
+                label="Schema"
+              >
+                <DataStructureListWrapper>
+                  <DataStructureList
+                    context={context}
+                    currentActionId={params?.queryId || ""}
+                    datasourceId={datasourceId || ""}
+                    datasourceStructure={datasourceStructure}
+                    pluginName={pluginName}
+                    step={0}
+                  />
+                </DataStructureListWrapper>
+              </Collapsible>
+            </CollapsibleSection>
+          )}
 
-      {showSchema && (
-        <CollapsibleSection
-          height={
-            datasourceStructure?.tables?.length && !isLoadingSchema
-              ? "50%"
-              : "auto"
-          }
-          id={SCHEMA_SECTION_ID}
-          ref={schemaRef}
-        >
-          <Collapsible
-            CustomLabelComponent={DatasourceStructureHeader}
-            containerRef={schemaRef}
-            datasourceId={datasourceId}
-            expand={!showSuggestedWidgets}
-            label="Schema"
-          >
-            <DataStructureListWrapper>
-              <DataStructureList
-                context={context}
-                currentActionId={params?.queryId || ""}
-                datasourceId={datasourceId || ""}
-                datasourceStructure={datasourceStructure}
-                pluginName={pluginName}
-                step={0}
+          {showSchema && <Divider />}
+          {showSuggestedWidgets ? (
+            <CollapsibleSection height={"40%"} marginTop={12}>
+              <SuggestedWidgets
+                actionName={actionName}
+                hasWidgets={hasWidgets}
+                suggestedWidgets={suggestedWidgets as SuggestedWidget[]}
               />
-            </DataStructureListWrapper>
-          </Collapsible>
-        </CollapsibleSection>
-      )}
-
-      {showSchema && <Divider />}
-      {showSuggestedWidgets ? (
-        <CollapsibleSection height={"40%"} marginTop={12}>
-          <SuggestedWidgets
-            actionName={actionName}
-            hasWidgets={hasWidgets}
-            suggestedWidgets={suggestedWidgets as SuggestedWidget[]}
-          />
-        </CollapsibleSection>
-      ) : (
-        <DisabledCollapsible
-          label={createMessage(BINDING_SECTION_LABEL)}
-          tooltipLabel={createMessage(BINDINGS_DISABLED_TOOLTIP)}
-        />
-      )}
+            </CollapsibleSection>
+          ) : (
+            <DisabledCollapsible
+              label={createMessage(BINDING_SECTION_LABEL)}
+              tooltipLabel={createMessage(BINDINGS_DISABLED_TOOLTIP)}
+            />
+          )}
+        </CollapsibleGroup>
+      </SectionWrapper>
     </SideBar>
   );
 }
